Moved parameters to dedicated package
